Fixed codeQL warning from regex overflow. Moved flattenTree method to common. Will be added to when future methods are created.
