+++ Tombolo/client-reactjs/src/components/admin/workunitHistory/common.js
@@ -1,0 +1,12 @@
+function flattenTree(nodes) {
+  const out = [];
+  const visit = arr =>
+    arr?.forEach(n => {
+      out.push(n);
+      if (n.children) visit(n.children);
+    });
+  visit(nodes);
+  return out;
+}
+
+export { flattenTree };
